requirements.txt
streamlit==1.44.1
tensorflow==2.13.0
numpy
opencv-python
pillow